Data.Filesystem.FunFS.API

Description

Module : Data.Filesystem.FunFS.API

Description : FunFS API.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Directory

An open directory handle. Composes Ext2.Directory and Handle

data DirEntry

A directory entry. Associates a node with the name given it by its parent directory.

getEntryName :: DirEntry -> String

Get the name associated with a directory entry.

getEntryNode :: DirEntry -> Node

Get the node associated with a directory entry.

openDir

Arguments

:: String
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as superuser.

openDirP

Arguments

:: Path
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as superuser.

openDirAs

Arguments

:: User
The user whose permissions will be used when opening the directory.

-> String
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as a given user.

openDirAsP

Arguments

:: User
The user whose permissions will be used when opening the directory.

-> Path
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as a given user.

closeDir :: Directory -> FS ()

Close a directory.

readDir :: Directory -> FS DirEntry

Read the next directory entry from a directory. It is an error to attempt to read more than the number of entries in the directory, which can be checked with getDirSize.

mapDir :: (DirEntry -> FS a) -> Directory -> FS [a]

Map a function over all the entries in a directory. NB: The reading position is not altered (unless the function argument to mapDir causes it) so thatthe code: entry1 <- readDir dir mapDir f dir entry2 <- readDir Has the same effect as if mapDir was removed, i.e. entry1 and entry2 are adjacent entries.

getDirHandle :: Directory -> FS Handle

Get the handle associated with a directory.

readDirToEnd :: Directory -> FS [DirEntry]

Read all the entries of a directory into a list.

getDirSize :: Directory -> FS (Maybe Int)

Get the number of entries in a directory.

data File

An open fileA handle to an opened regular file.

fOpen

Arguments

:: String
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as superuser.

fOpenP

Arguments

:: Path
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as superuser.

fOpenAs

Arguments

:: User
The user whose permissions will be used when opening the file.

-> String
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as a given user.

fOpenAsP

Arguments

:: User
The user whose permissions will be used when opening the file.

-> Path
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as a given user.

fClose :: File -> FS ()

Close a file.

fSeek :: Seek -> Int -> File -> FS ()

Change the read/write position of a file.

fRead

Arguments

:: File
The file to read from.

-> Int
How many bytes to read.

-> FS ByteString
The resulting ByteString is returned in FS, which is updated.

Read from a file.

fGetBytes

Arguments

:: File
The file to read from.

-> Int
How many bytes to read.

-> FS [Word8]
The resulting list is returned in FS, which is updated.

Read bytes from a file.

fGetHandle :: File -> FS Handle

Get the handle associated with a file.

fGetSize :: File -> FS (Maybe Int)

Get the number of entries in a file.

fGetPosition :: File -> FS (Maybe Int)

Get the read/write position of a file.

data Filesystem

Filesystem.

Constructors

NotMounted
An invalid/unmounted filesystem. | A mounted filesystem.

Filesystem

Fields

volume :: Volume

Ext2 volume.

handles :: [Maybe HandleInfo]

List of handles.

rootDirectory :: Directory

The root directory.

type FS a = State Filesystem a

FS monad.

get :: Monad m => StateT s m s

Fetch the current value of the state within the monad.

put :: Monad m => s -> StateT s m ()

put s sets the state within the monad to s.

sync :: FS a -> Filesystem -> IO a

Apply a filesystem action to a mounted filesystem and return the result in IO. Due to lazy evaluation, filesystem operations are not evaluated until sync is called.

appendFDHandleInfo :: HandleInfo -> FS Int

Append the HandleInfo for a newly opened handle to a filesystem's list of open handles and get a file descriptor unique to it. It is NOT an error if there is already a handle to the same node, though it may cause strange behaviour.

updateFDHandleInfo :: Int -> (Maybe HandleInfo -> Maybe HandleInfo) -> FS (Maybe HandleInfo)

Update the HandleInfo referenced by a file descriptor by applying a function to it and return the result. It is an error if the file descriptor refers to a closed or otherwise invalid handle.

isInUse :: FS Bool

Indicates whether the filesystem is in use, i.e. if there are open handles.

getStatistics :: FS String

Get printable statistics about the filesystem.

openHandles :: Filesystem -> [HandleInfo]

Get a list of currently open handles.

getFDHandleInfo :: Int -> FS (Maybe HandleInfo)

Get the HandleInfo referenced by a file descriptor. If the handle is closed, the result will be Nothing.

haveFDHandleInfo :: Int -> FS Bool

Indicates whether a file descriptor is associated with an open handle.

maybeHI :: a -> (HandleInfo -> FS a) -> Int -> FS a

Helper for functions which execute conditionally on whether a handle is opened.

data Handle

A Filesystem handle - references an open file or directory within the filesystem. Essentially equivalent to POSIX - file descriptors (in fact it is exactly equal).

openH

Arguments

:: User
Use the permissions of a user to open the file.

-> Node
The node to which the handle should refer.

-> OpenMode
The opening mode.

-> FS Handle
An open handle to the node wrapped in updated FS.

Open a handle to a filesystem node.

openHWithSize :: User -> Node -> OpenMode -> Int -> FS Handle

Open a handle, overriding the size reported by the node. This is a workaround to allow directories to report their size as the number of entries they contain, rather than the sum of the sizes of the entries.

closeH :: Handle -> FS ()

Close a handle.

seekH :: Seek -> Int -> Handle -> FS ()

Update the read/write position of a handle.

readH :: Handle -> Int -> FS ByteString

Read data from a handle.

isHOpen :: Handle -> FS Bool

Indicates whether the handle is open.

isHClosed :: Handle -> FS Bool

Indicates whether the handle is closed.

getHInfo :: Handle -> FS (Maybe HandleInfo)

Get the HandleInfo associated with a Handle. If the handle is closed, Nothing is returned.

getHType :: Handle -> FS (Maybe NodeType)

Get the NodeType associated with a Handle. If the handle is closed, Nothing is returned.

getHNode :: Handle -> FS (Maybe Node)

Get the Node associated with a Handle. If the handle is closed, Nothing is returned.

getHOpenMode :: Handle -> FS (Maybe OpenMode)

Get the OpenMode associated with a Handle. If the handle is closed, Nothing is returned.

getHPermissions :: Handle -> FS (Maybe Permission)

Get the Permission associated with a Handle. If the handle is closed, Nothing is returned.

getHPosition :: Handle -> FS (Maybe Int)

Get the position associated with a Handle. If the handle is closed, Nothing is returned.

getHSize :: Handle -> FS (Maybe Int)

Get the size associated with a Handle. If the handle is closed, Nothing is returned.

data HandleInfo

Describes an open or directory.

Constructors

HandleInfo

Fields

hNode :: Node

The filesystem node.

hType :: NodeType

The node type.

hOpenMode :: OpenMode

The opening mode.

hPermissions :: Permission

The permissions of the handle.

hPosition :: Int

The read/write position.

hSize :: Int

The size of the file/directory in bytes.

makeHandleInfo :: User -> Node -> OpenMode -> HandleInfo

Create and populate a HandleInfo record. Also performs permissions checks.

mount

Arguments

:: String
A path to the file or device *in the *host* FS* which contains a filesystem.

-> IO Filesystem
The filesystem within IO. Use sync to use the filesystem in FS.

Mount a filesystem.

unmount :: FS ()

Unmount a filesystem. The unmount will fail if there are open handles.

forceUnmount :: FS ()

Unmount a filesystem forcibly. Invalidates all open handles.

withMount :: String -> FS a -> IO a

Mount a filesystem and apply a filesystem action to it, returning the result in IO.

data NodeType

Node type.

Constructors

FileNode
A regular file.

DirNode
A directory. Contains references to other nodes.

SpecialNode
A special node such as a device or symbolic link.

isFileNode :: NodeType -> Bool

Indicates whether the node type is FileNode.

isDirNode :: NodeType -> Bool

Indicates whether the node type is DirectNode.

isSpecialNode :: NodeType -> Bool

Indicates whether the node type is SpecialNode.

data OpenMode

File open mode.

Constructors

ReadOnly
Allow reading only.

WriteOnly
Allow writing only.

ReadWrite
Allow reading and writing.

Append
Allow writing at the end of the file only.

data Path

Path .

elems :: Path -> [String]

sep :: Char

The path separator.

makePath :: String -> Path

Make a path from a string.

buildPath :: [String] -> Path

Build a path from separate elements.

root :: Path -> String

Get the root element of a path.

tree :: Path -> Path

Get all the elements of a path except the root.

dirs :: Path -> Path

Get all the elements of a path except the leaf.

leaf :: Path -> String

Get the leaf element of a path.

parent :: Path -> String

Get the parent of the leaf element of a path.

isEmpty :: Path -> Bool

Indicates whether a path is empty.

isRoot :: Path -> Bool

isLeaf :: Path -> Bool

isTree :: Path -> Bool

data Permission

File permission.

Constructors

Permission

Fields

readable :: Bool

Whether the file is readable.

writeable :: Bool

Whether the file is writeable.

executable :: Bool

Whether the file is executable.

getPermissions :: User -> Node -> Permission

Get permissions of a node wrt a given user. In other words, tests whether a user is allowed to read, write and - execute a node and wraps this in a Permission.

data Seek

Seek type.

Constructors

Begin
Relative to file beginning.

End
Relative to file end.

CurPos
Relative to current position.

isBegin :: Seek -> Bool

Indicates whether to seek from the beginning.

isEnd :: Seek -> Bool

Indicates whether to seek from the end.

isCurPos :: Seek -> Bool

Indicates whether to seek from the current position.

computeSeek

Arguments

:: Integral a

=> Seek
The seek type

-> (a, a, a)
The seek boundaries in form (minimum,current,maximum)

-> a
The proposed seek offset. For example 1 if seeking by 1 byte.

-> a
The absolute position to seek to.

Computes a new seek position bounded by the minimum, current and maximum values.

data User

User.

Constructors

User

Fields

userID :: Int

UID of user.

groups :: [Int]

GIDs of groups user belongs to.

superuser :: User

The superuser, aka root.