Data.FileSystem.FunFS.API.Directory

Description

Module : Data.FileSystem.FunFS.API.Directory

Description : Directory API.

Copyright : (c) 2017-2018 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Directory

An open directory handle. Composes Ext2.Directory and Handle

data DirEntry

A directory entry. Associates a node with the name given it by its parent directory.

getEntryName :: DirEntry -> String

Get the name associated with a directory entry.

getEntryNode :: DirEntry -> Node

Get the node associated with a directory entry.

openDir

Arguments

:: String
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as superuser.

openDirP

Arguments

:: Path
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as superuser.

openDirAs

Arguments

:: User
The user whose permissions will be used when opening the directory.

-> String
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as a given user.

openDirAsP

Arguments

:: User
The user whose permissions will be used when opening the directory.

-> Path
A path which names a directory.

-> OpenMode
The directory open mode (read/write).

-> FS Directory
The opened directory in FS.

Open a directory as a given user.

closeDir :: Directory -> FS ()

Close a directory.

readDir :: Directory -> FS DirEntry

Read the next directory entry from a directory. It is an error to attempt to read more than the number of entries in the directory, which can be checked with getDirSize.

mapDir :: (DirEntry -> FS a) -> Directory -> FS [a]

Map a function over all the entries in a directory. NB: The reading position is not altered (unless the function argument to mapDir causes it) so thatthe code: entry1 <- readDir dir mapDir f dir entry2 <- readDir Has the same effect as if mapDir was removed, i.e. entry1 and entry2 are adjacent entries.

getDirHandle :: Directory -> FS Handle

Get the handle associated with a directory.

readDirToEnd :: Directory -> FS [DirEntry]

Read all the entries of a directory into a list.

getDirSize :: Directory -> FS (Maybe Int)

Get the number of entries in a directory.