Data.Filesystem.FunFS.API.File

Description

Module : Data.Filesystem.FunFS.API.File

Description : File API.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data File

An open fileA handle to an opened regular file.

fOpen

Arguments

:: String
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as superuser.

fOpenP

Arguments

:: Path
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as superuser.

fOpenAs

Arguments

:: User
The user whose permissions will be used when opening the file.

-> String
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as a given user.

fOpenAsP

Arguments

:: User
The user whose permissions will be used when opening the file.

-> Path
A path which names a file.

-> OpenMode
The file open mode (read/write).

-> FS File
The opened file in FS.

Open a file as a given user.

fClose :: File -> FS ()

Close a file.

fSeek :: Seek -> Int -> File -> FS ()

Change the read/write position of a file.

fRead

Arguments

:: File
The file to read from.

-> Int
How many bytes to read.

-> FS ByteString
The resulting ByteString is returned in FS, which is updated.

Read from a file.

fGetBytes

Arguments

:: File
The file to read from.

-> Int
How many bytes to read.

-> FS [Word8]
The resulting list is returned in FS, which is updated.

Read bytes from a file.

fGetHandle :: File -> FS Handle

Get the handle associated with a file.

fGetSize :: File -> FS (Maybe Int)

Get the number of entries in a file.

fGetPosition :: File -> FS (Maybe Int)

Get the read/write position of a file.