Data.FileSystem.FunFS.API.FileSystem

Description

Module : Data.FileSystem.FunFS.API.FileSystem

Description : FileSystem state and FS monad.

Copyright : (c) 2017-2018 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data FileSystem

FileSystem.

Constructors

NotMounted
An invalid/unmounted filesystem. | A mounted filesystem.

FileSystem

Fields

volume :: Volume

Ext2 volume.

handles :: [Maybe HandleInfo]

List of handles.

rootDirectory :: Directory

The root directory.

type FS a = State FileSystem a

FS monad.

get :: Monad m => StateT s m s

put :: Monad m => s -> StateT s m ()

sync :: FS a -> FileSystem -> IO a

Apply a filesystem action to a mounted filesystem and return the result in IO. Due to lazy evaluation, filesystem operations are not evaluated until sync is called.

appendFDHandleInfo :: HandleInfo -> FS Int

Append the HandleInfo for a newly opened handle to a filesystem's list of open handles and get a file descriptor unique to it. It is NOT an error if there is already a handle to the same node, though it may cause strange behaviour.

updateFDHandleInfo :: Int -> (Maybe HandleInfo -> Maybe HandleInfo) -> FS (Maybe HandleInfo)

Update the HandleInfo referenced by a file descriptor by applying a function to it and return the result. It is an error if the file descriptor refers to a closed or otherwise invalid handle.

isInUse :: FS Bool

Indicates whether the filesystem is in use, i.e. if there are open handles.

getStatistics :: FS String

Get printable statistics about the filesystem.

openHandles :: FileSystem -> [HandleInfo]

Get a list of currently open handles.

getFDHandleInfo :: Int -> FS (Maybe HandleInfo)

Get the HandleInfo referenced by a file descriptor. If the handle is closed, the result will be Nothing.

haveFDHandleInfo :: Int -> FS Bool

Indicates whether a file descriptor is associated with an open handle.

maybeHI :: a -> (HandleInfo -> FS a) -> Int -> FS a

Helper for functions which execute conditionally on whether a handle is opened.