Data.Filesystem.FunFS.API.Handle

Description

Module : Data.Filesystem.FunFS.API.Handle

Description : Filesystem I/O handle.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Handle

A Filesystem handle - references an open file or directory within the filesystem. Essentially equivalent to POSIX - file descriptors (in fact it is exactly equal).

openH

Arguments

:: User
Use the permissions of a user to open the file.

-> Node
The node to which the handle should refer.

-> OpenMode
The opening mode.

-> FS Handle
An open handle to the node wrapped in updated FS.

Open a handle to a filesystem node.

openHWithSize :: User -> Node -> OpenMode -> Int -> FS Handle

Open a handle, overriding the size reported by the node. This is a workaround to allow directories to report their size as the number of entries they contain, rather than the sum of the sizes of the entries.

closeH :: Handle -> FS ()

Close a handle.

seekH :: Seek -> Int -> Handle -> FS ()

Update the read/write position of a handle.

readH :: Handle -> Int -> FS ByteString

Read data from a handle.

isHOpen :: Handle -> FS Bool

Indicates whether the handle is open.

isHClosed :: Handle -> FS Bool

Indicates whether the handle is closed.

getHInfo :: Handle -> FS (Maybe HandleInfo)

Get the HandleInfo associated with a Handle. If the handle is closed, Nothing is returned.

getHType :: Handle -> FS (Maybe NodeType)

Get the NodeType associated with a Handle. If the handle is closed, Nothing is returned.

getHNode :: Handle -> FS (Maybe Node)

Get the Node associated with a Handle. If the handle is closed, Nothing is returned.

getHOpenMode :: Handle -> FS (Maybe OpenMode)

Get the OpenMode associated with a Handle. If the handle is closed, Nothing is returned.

getHPermissions :: Handle -> FS (Maybe Permission)

Get the Permission associated with a Handle. If the handle is closed, Nothing is returned.

getHPosition :: Handle -> FS (Maybe Int)

Get the position associated with a Handle. If the handle is closed, Nothing is returned.

getHSize :: Handle -> FS (Maybe Int)

Get the size associated with a Handle. If the handle is closed, Nothing is returned.