Data.FileSystem.FunFS.API.Mount

Description

Module : Data.FileSystem.FunFS.API.Mount

Description : Mount/umount filesystem.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

mount

Arguments

:: String
A path to the file or device *in the *host* FS* which contains a filesystem.

-> IO FileSystem
The filesystem within IO. Use sync to use the filesystem in FS.

Mount a filesystem.

unmount :: FS ()

Unmount a filesystem. The unmount will fail if there are open handles.

forceUnmount :: FS ()

Unmount a filesystem forcibly. Invalidates all open handles.

withMount :: String -> FS a -> IO a

Mount a filesystem and apply a filesystem action to it, returning the result in IO.