Data.FileSystem.FunFS.API.Permission

Description

Module : Data.FileSystem.FunFS.API.Permission

Description : File permissions.

Copyright : (c) 2017-2018 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Permission

File permission.

Constructors

Permission

Fields

readable :: Bool

Whether the file is readable.

writeable :: Bool

Whether the file is writeable.

executable :: Bool

Whether the file is executable.

getPermissions :: User -> Node -> Permission

Get permissions of a node wrt a given user. In other words, tests whether a user is allowed to read, write and - execute a node and wraps this in a Permission.