Data.Filesystem.FunFS.Ext2.Conf

Description

Module : Data.Filesystem.FunFS.Ext2.Conf

Description : Ext2 configuration.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Conf

Ext2 configuration. This can be used to:

1. Create a superblock when making a new filesystem.

2. Update a superblock when tuning an existing filesystem.

3. Return a high-level extract of the superblock when mounting a filesystem.

Constructors

Conf

Fields

volumeSize :: Bytes

The size of the volume.

blockSize :: Bytes

The block size in bytes.

blocksPerGroup :: Int

The number of blocks per group.

bytesPerInode :: Bytes

The number of bytes of disk space to allow per inode.

maxMountCount :: Int

The maximum number of mounts before consistency check required.

rBlocksCount :: Int

The number of reserved blocks.

checkInterval :: Int

The interval in seconds between filesystem checks (0=never).

volumeName :: String

The volume label (16 characters).

minimumVolumeSize :: Bytes

defaultBlockSize :: Bytes

defaultBytesPerInode :: Bytes

defaultRBlocksCount :: Int

defaultCheckInterval :: Int

defaultVolumeName :: String

defaultConf :: Conf

Default configuration.