Data.FileSystem.FunFS.Ext2.Directory

Description

Module : Data.FileSystem.FunFS.Ext2.Directory

Description : Directories.

Copyright : (c) 2017-2018 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Directory

Directory - a node which contains other nodes.

makeDirectory :: Node -> Directory

Make a directory from a node. The node must be of type Directory.

getRootDirectory :: Volume -> Directory

Get the root directory.

getEntries :: Directory -> [DirEntry]

getNames :: Directory -> [String]

Get a list of the names of a directory's children.

getChildren :: Directory -> [Node]

Get a list of a directory's children (subordinate nodes).

getChildByName :: Directory -> String -> Maybe Node

Get the child node named by s in dir if it exists or Nothing if it does not.

getChildByPath :: Directory -> Path -> Node

Scan the directory tree for a node named by path. It is an error if the node does not exist. Paths are assumed to be relative from the directory given. An empty path or one containing only "." will match the given directory.