Data.Filesystem.FunFS.Ext2.LowLevel.Block

Description

Module : Data.Filesystem.FunFS.Ext2.LowLevel.Block

Description : Primitives for handling block-based I/O.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Block

A block of data.

bytes :: Block -> ByteString

size :: Block -> Int64

makeBlock :: ByteString -> Int64 -> Block

Make a Block from a ByteString.

makeBlockList :: ByteString -> Int64 -> [Block]

Divide a ByteString into a list of Blocks.

peekObject :: Serialisable a => Block -> Int64 -> a

Decode a Serialisable object starting at index i in a block. The entire object must fit within the block.

peekObjects :: (Serialisable a, SizeOf a) => Block -> Int64 -> [a]

Create an infinite list of Serialisable objects starting at index i in a block.

pokeObject :: Serialisable a => Block -> Int64 -> a -> Block

Encode a Serialisable object to index i in a block. The entire object must fit within the block. The original contents of the block (from i to the end of the object) are overwritten but the size of the block remains constant.

peek :: Block -> Int64 -> Word8

Extract a single byte from a block at a given index.

poke :: Block -> Int64 -> Word8 -> Block

Insert a single byte into a block at a given index. The original byte is overwritten but the size of the block remains constant.