Data.FileSystem.FunFS.Ext2.LowLevel.Struct

Description

Module : Data.FileSystem.FunFS.Ext2.LowLevel.Struct

Description : Serialisable records for representing low-level data structures.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Field

Structure field. This can be a scalar field, which represents a single value, or a vector field, which represents an array.

field :: String -> TypeQ -> Field

Create a field which is a single value.

array :: String -> TypeQ -> Int -> Field

Create a field which is an array of values.

struct :: String -> [Field] -> [String] -> DecsQ

Generate a serialisable record.

module Data.FileSystem.FunFS.Util.SizeOf

module Data.FileSystem.FunFS.Util.Serialisable

module Data.FileSystem.FunFS.Util.Strings