Data.FileSystem.FunFS.Ext2.LowLevel.SuperBlock

Description

Module : Data.FileSystem.FunFS.Ext2.LowLevel.SuperBlock

Description : Superblock

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data SuperBlock

Superblock.

Constructors

SuperBlock

Fields

sInodesCount :: Word32

sBlocksCount :: Word32

sRBlocksCount :: Word32

sFreeBlocksCount :: Word32

sFreeInodesCount :: Word32

sFirstDataBlock :: Word32

sLogBlockSize :: Word32

sLogFragSize :: Word32

sBlocksPerGroup :: Word32

sFragsPerGroup :: Word32

sInodesPerGroup :: Word32

sMtime :: Word32

sWtime :: Word32

sMountCount :: Word16

sMaxMountCount :: Word16

sMagic :: Word16

sState :: Word16

sErrors :: Word16

sMinorRevLevel :: Word16

sLastCheck :: Word32

sCheckInterval :: Word32

sCreatorOs :: Word32

sRevLevel :: Word32

sDefResUid :: Word16

sDefResGid :: Word16

sFirstIno :: Word32

sInodeSize :: Word16

sBlockGroupNr :: Word16

sFeatureCompat :: Word32

sFeatureIncompat :: Word32

sFeatureRoCompat :: Word32

sUuid :: [Word8]

sVolumeName :: [Word8]

sLastMounted :: [Word8]

sAlgoBitmap :: Word32

sPreallocBlocks :: Word8

sPreallocDirBlocks :: Word8

alignment :: Word16

sJournalUuid :: [Word8]

sJournalInum :: Word32

sJournalDev :: Word32

sLastOrphan :: Word32

sHashSeed :: [Word32]

sDefHashVersion :: Word8

padding :: [Word8]

sDefaultMountOptions :: Word32

sFirstMetaBg :: Word32

data State

FileSystem state.

Constructors

ValidFS
FileSystem cleanly unmounted.

ErrorFS
FileSystem may contain errors.

data Errors

Value indicating what the

Constructors

Continue
Continue as if nothing happened.

RO
Remount as read-only.

Panic
Trigger kernel panic.

superBlockMagic :: Word16

The superblock magic number.

makeSuperBlock :: String -> Word32 -> Word32 -> Word32 -> Word32 -> Word16 -> Word32 -> Word32 -> IO SuperBlock

defaultSuperBlock :: SuperBlock

A basic superblock with sensible default values for some members.

Can be used as the basis for a new superblock when creating a filesystem.

NB: The following members are undefined and must be updated:

sInodesCount

sBlocksCount

sRBlocksCount

sFreeBlocksCount

sFreeInodesCount

sMtime

sWtime

sLastCheck

sCheckInterval

sUuid

sVolumeName

NB: sFirstDataBlock should be updated if the block size is changed. Its value is 0 if the block size is greater than

1024. If the block size is 1024, then sFirstDataBlock must have the value 1.

checkSuperBlock :: SuperBlock -> SuperBlock

Check whether the superblock is valid. If it is, the superblock is returned. If not, an error string describing the first problem detected is returned.

readSuperBlock :: ByteString -> SuperBlock

Try to decode a superblock from a bytestring.

NB: The superblock always starts at the 1024th byte and is 1024 bytes long. Therefore the bytestring must be at least 2048 bytes long. This is not checked for performance reasons.

If the superblock magic number is correct, then the superblock is returned. If the magic number is incorrect, the superblock is either not valid or has been corrupted, and Nothing is returned. In this case, the filesystem may be recoverable using fsck(8).

getBlockSize :: SuperBlock -> Int

Get the block size of a filesystem. This requires extra computation from the sLogBlockSize field of the superblock.

filesystemSize :: SuperBlock -> Int

Compute the size of the filesystem.

countBlockGroups :: SuperBlock -> Int

Compute the number of block groups in the filesystem.

defaultLogBlockSize :: Word32

defaultBlocksPerGroup :: Word32

defaultMaxMountCount :: Word16