Data.FileSystem.FunFS.Ext2.Node

Description

Module : Data.FileSystem.FunFS.Ext2.Node

Description : I/O operations on node.

Copyright : (c) 2017-2018 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Node

A filesystem node. This is a thin wrapper around inode.

data Inode

inum :: Node -> Int

inode :: Node -> Inode

volume :: Node -> Volume

rootNodeNumber :: Int

The root directory is always inode #2.

getRootNode :: Volume -> Node

The root node.

getNodeByInum :: Volume -> Int -> Node

isRootNode :: Node -> Bool

Indicates whether a node is the filesystem root node. The root node is a directory with inode number 2 and no parent node.

nodeType :: Node -> NodeType

Get a node's type.

nodeSize :: Integral a => Node -> a

Get the size in bytes of a node.

isFile :: Node -> Bool

Indicates whether a node is a file.

isDirectory :: Node -> Bool

Indicates whether a node is a directory.

isSpecial :: Node -> Bool

Indicates whether a node is a special.

isOwner :: User -> Node -> Bool

Indicates whether a user owns a file. Always returns true if the user is root.

isInGroup :: User -> Node -> Bool

Indicates whether a user is in the group which owns a file. Always returns true if the user is root.

isReadable :: User -> Node -> Bool

Indicates whether a node can be opened for reading.

isWriteable :: User -> Node -> Bool

Indicates whether a node can be opened for writing.

isExecutable :: User -> Node -> Bool

Indicates whether a node can be opened for execution.

getBlocks :: Node -> [Block]

Get a list of the blocks being used by a node.

getBytes :: Node -> ByteString

Get a ByteString containing the entire contents of a node.

getBytes' :: Node -> (Int64, Int64) -> ByteString

Get a ByteString containing a section of a node's data.

getObject :: Serialisable a => Node -> (Int64, Int64) -> a

Decode an object of given size starting at a given offset within a node.