Data.FileSystem.FunFS.Ext2.Path

Description

Module : Data.FileSystem.FunFS.Ext2.Path

Description : Delimited file path.

Copyright : (c) 2017 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

data Path

Path .

elems :: Path -> [String]

sep :: Char

The path separator.

makePath :: String -> Path

Make a path from a string.

buildPath :: [String] -> Path

Build a path from separate elements.

root :: Path -> String

Get the root element of a path.

tree :: Path -> Path

Get all the elements of a path except the root.

dirs :: Path -> Path

Get all the elements of a path except the leaf.

leaf :: Path -> String

Get the leaf element of a path.

parent :: Path -> String

Get the parent of the leaf element of a path.

isEmpty :: Path -> Bool

Indicates whether a path is empty.

isRoot :: Path -> Bool

isLeaf :: Path -> Bool

isTree :: Path -> Bool