Data.FileSystem.FunFS.Util.Serialisable

Description

Module : Data.FileSystem.FunFS.Util.Serialisable

Description : Serialisable typeclass.

Copyright : (c) 2017-2018 Chris Swinchatt

License : MIT

Maintainer : Chris Swinchatt c.swinchatt1@uni.brighton.ac.uk

Stability : experimental

Portability : portable

Documentation

class Serialisable a where

Typeclass implemented by serialisable objects. - Minimum complete definition: get and put. - - Note: We cannot use the Data.Binary.Binary class here because it always uses network byte ordering (big endian), - while Ext2 always uses little endian.

Minimal complete definition

get, put

Methods

get :: Get a

Retrieve a value of type a.

put :: a -> Put

Store a value of type a.

getList :: Int -> Get [a]

Retrieve a list of as.

putList :: [a] -> Put

Store a list of as.

decode :: ByteString -> a

Decode a value of type a from a lazy ByteString.

encode :: a -> ByteString

Encode a value of type a into a lazy ByteString.

decodeList :: Int -> ByteString -> [a]

Decode a list of as from a lazy ByteString.

encodeList :: [a] -> ByteString

Encode a list of as into a lazy ByteString.